print/requirements.txt
attrs==21.2.0
brother-ql==0.9.4
click==8.0.3
future==0.18.2
packbits==0.6
Pillow==8.4.0
pyusb==1.2.1
